moved delete profile button and removed password confirm from signup page
