+++ client/src/components/Profile.js
@@ -1,5 +1,8 @@
 import { useQuery } from "@apollo/client";
 import { QUERY_ME } from "../utils/queries";
+import { useMutation } from "@apollo/client";
+import auth from "../utils/auth";
+import { DELETE_PROFILE } from "../utils/mutations";
 import { Link, useHref } from "react-router-dom";
 
 import React, { useEffect, useState } from "react";
@@ -7,9 +10,14 @@ import { LOGIN_PROFILE } from "../utils/mutations";
 
 function Profile() {
   const [userData, setUserData] = useState({});
+  const [removeProfile, { error: deleteProfileErr, data: deleteProfileData }] =
+    useMutation(DELETE_PROFILE);
   const { loading, data, error } = useQuery(QUERY_ME);
   if (error) {
     console.log(JSON.stringify(error));
+  }
+  if (deleteProfileErr) {
+    console.log(JSON.stringify(deleteProfileErr));
   }
 
   useEffect(() => {
@@ -18,6 +26,19 @@ function Profile() {
       console.log(data);
     }
   }, [loading, data]);
+
+  async function handleDelete(e) {
+    try {
+      const { data } = await removeProfile({
+        variables: {},
+      });
+      auth.logout();
+
+      console.log(data);
+    } catch (error) {
+      console.log(JSON.stringify(error));
+    }
+  }
 
   return (
     <>
@@ -41,6 +62,14 @@ function Profile() {
                   Add a Child
                 </a>
               </button>
+              <Link
+                type="button"
+                onClick={handleDelete}
+                className="btn btn-danger btn-lg"
+                to="/"
+              >
+                Delete profile
+              </Link>
             </div>
           </div>
           <button className="dkgreen-bkg text-white">
